+++ src/main/java/org/sonar/plugins/delphi/type/DelphiHelperType.java
@@ -3,10 +3,12 @@ package org.sonar.plugins.delphi.type;
 import static org.sonar.plugins.delphi.type.StructKind.CLASS_HELPER;
 import static org.sonar.plugins.delphi.type.StructKind.RECORD_HELPER;
 
+import java.util.List;
 import java.util.Set;
 import org.jetbrains.annotations.NotNull;
 import org.sonar.plugins.delphi.antlr.ast.node.ClassHelperTypeNode;
 import org.sonar.plugins.delphi.antlr.ast.node.HelperTypeNode;
+import org.sonar.plugins.delphi.antlr.ast.node.TypeDeclarationNode;
 import org.sonar.plugins.delphi.symbol.scope.DelphiScope;
 import org.sonar.plugins.delphi.type.Type.HelperType;
 
@@ -14,16 +16,21 @@ public class DelphiHelperType extends DelphiStructType implements HelperType {
   private final Type extendedType;
 
   private DelphiHelperType(
-      String image, DelphiScope scope, Set<Type> parents, Type extendedType, StructKind kind) {
-    super(image, scope, parents, kind);
+      List<ImagePart> imageParts,
+      DelphiScope scope,
+      Set<Type> parents,
+      Type extendedType,
+      StructKind kind) {
+    super(imageParts, scope, parents, kind);
     this.extendedType = extendedType;
   }
 
   public static HelperType from(HelperTypeNode node) {
+    TypeDeclarationNode declaration = (TypeDeclarationNode) node.jjtGetParent();
     StructKind kind = (node instanceof ClassHelperTypeNode) ? CLASS_HELPER : RECORD_HELPER;
 
     return new DelphiHelperType(
-        node.jjtGetParent().getImage(),
+        createImageParts(declaration),
         node.getScope(),
         node.getParentTypes(),
         node.getFor().getType(),