+++ src/main/java/org/sonar/plugins/delphi/type/Type.java
@@ -8,6 +8,7 @@ import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.sonar.plugins.delphi.symbol.scope.DelphiScope;
+import org.sonar.plugins.delphi.type.generic.TypeSpecializationContext;
 
 public interface Type {
 
@@ -35,11 +36,22 @@ public interface Type {
   Set<Type> parents();
 
   /**
+   * Returns whether this type can be specialized.
+   *
+   * <p>NOTE: When specializing a type, this method does not need to be called. If a Type cannot be
+   * specialized, then <code>Type.specialize</code> will return <code>this</code>.
+   *
+   * @param context a specialization context, contains information to help specialize the type
+   * @return true if this type can be specialized
+   */
+  boolean canBeSpecialized(TypeSpecializationContext context);
+
+  /**
    * If this is a generic type, then we can specialize the type with some given type parameters and
    * arguments.
    *
    * @param context a specialization context, contains information to help specialize the type
-   * @return the specialized type, or unknown type if generic specialization fails
+   * @return the specialized type, or <code>this</code> if generic specialization wasn't possible
    */
   Type specialize(TypeSpecializationContext context);