+++ src/test/resources/org/sonar/plugins/delphi/symbol/generics/ArrayAssignmentCompatibility.pas
@@ -1,4 +1,4 @@
-unit AssignmentCompatibility;
+unit ArrayAssignmentCompatibility;
 
 {This is a sample Delphi file.}